Add redfractor changes for frontend
