+++ frontend/.eslintrc.js
@@ -24,5 +24,7 @@ module.exports = {
     'arrow-body-style': 0,
     'react/jsx-props-no-spreading': 0,
     'no-restricted-syntax': 0,
+    'jsx-a11y/click-events-have-key-events': 0,
+    'jsx-a11y/no-static-element-interactions': 0,
   },
 };
