+++ frontend/src/Components/index.js
@@ -2,6 +2,3 @@ export { default as Button } from '../Material-UI/Components/Button/Button';
 export { default as Input } from '../Material-UI/Components/Input/Input';
 export { default as Select } from '../Material-UI/Components/Select/Select';
 export { default as NavBar } from './Navbar/NavBar';
-export { default as AccountDetails } from './AccountDetails/AccountDetails';
-export { default as Header } from './Header/Header';
-export { default as LandingBody } from './LandingBody/LandingBody';
