+++ frontend/src/Containers/App/App.css
@@ -5,12 +5,12 @@
 }
 
 .anchor {
-  color: var(--primary-dark);
+  color: var(--primary-color);
   cursor: pointer;
 }
 
 a {
-  color: var(--primary-dark);
+  color: var(--primary-color);
   cursor: pointer;
   text-decoration: none;
 }
