+++ frontend/src/Containers/Onboarding/Signup/SignupPage.js
@@ -41,5 +41,6 @@ const SignUp = () => {
       </form>
     </div>
   );
-}
+};
+
 export default SignUp;
