+++ frontend/src/Material-UI/import/index.js
@@ -1,21 +1,20 @@
-export { createMuiTheme } from '@material-ui/core/styles';
 export { ThemeProvider } from '@material-ui/styles';
 export { makeStyles, withStyles } from '@material-ui/core/styles';
 export { theme } from '../theme';
 export {
-    Button,
-    Container,
-    CssBaseline,
-    Typography,
-    AppBar,
-    Toolbar,
-    TextField,
-    Avatar,
-    Select,
-    FormControl,
-    InputLabel,
-    MenuItem,
-    Menu,
-    ListItemIcon,
-    ListItemText,
+  Button,
+  Container,
+  CssBaseline,
+  Typography,
+  AppBar,
+  Toolbar,
+  TextField,
+  Avatar,
+  Select,
+  FormControl,
+  InputLabel,
+  MenuItem,
+  Menu,
+  ListItemIcon,
+  ListItemText,
 } from '@material-ui/core';
